fixing syntax errors halfway

+++ src/assemble.h
@@ -12,5 +12,5 @@
 
 typedef struct {
     const char* instrname;
-    uint32_t (*handler)(char *, char* [], int);
+    uint32_t (*handler)(char *, char* [], size_t);
 } instr_dispatch;
